Move pre sortie data fetching into a class
